unfinished

+++ src/main/java/com/lais/valuegenerators/TimeValueGenerator.java
@@ -5,5 +5,5 @@ package com.lais.valuegenerators;
  */
 public interface TimeValueGenerator {
 
-    int generateValue();
+    long generateValue();
 }
